requirements.txt
flask
faiss-cpu
sentence-transformers
google-generativeai
numpy
gunicorn
